requirements.txt
Django==2.2
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
psycopg2-binary==2.8.4
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
